hiti/requirements.txt
Pillow>=9.0.0
cffi>=1.15.0
pywin32>=305; sys_platform == 'win32'
setuptools>=60.0.0
wheel>=0.37.0
pytest>=7.0.0